Create specialMentions package
