+++ app/controllers/specialMentions/AddSpecialMentionController.scala
@@ -14,33 +14,33 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.specialMentions
 
 import controllers.actions._
-import forms.SpecialMentionAdditionalInfoFormProvider
+import forms.AddSpecialMentionFormProvider
 import javax.inject.Inject
 import models.{LocalReferenceNumber, Mode}
 import navigation.Navigator
 import navigation.annotations.PreTaskListDetails
-import pages.SpecialMentionAdditionalInfoPage
+import pages.AddSpecialMentionPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.libs.json.Json
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import renderer.Renderer
 import repositories.SessionRepository
 import uk.gov.hmrc.play.bootstrap.controller.FrontendBaseController
-import uk.gov.hmrc.viewmodels.NunjucksSupport
+import uk.gov.hmrc.viewmodels.{NunjucksSupport, Radios}
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class SpecialMentionAdditionalInfoController @Inject()(
+class AddSpecialMentionController @Inject()(
   override val messagesApi: MessagesApi,
   sessionRepository: SessionRepository,
   @PreTaskListDetails navigator: Navigator,
   identify: IdentifierAction,
   getData: DataRetrievalActionProvider,
   requireData: DataRequiredAction,
-  formProvider: SpecialMentionAdditionalInfoFormProvider,
+  formProvider: AddSpecialMentionFormProvider,
   val controllerComponents: MessagesControllerComponents,
   renderer: Renderer
 )(implicit ec: ExecutionContext)
@@ -49,19 +49,20 @@ class SpecialMentionAdditionalInfoController @Inject()(
     with NunjucksSupport {
 
   private val form     = formProvider()
-  private val template = "specialMentionAdditionalInfo.njk"
+  private val template = "addSpecialMention.njk"
 
   def onPageLoad(lrn: LocalReferenceNumber, mode: Mode): Action[AnyContent] = (identify andThen getData(lrn) andThen requireData).async {
     implicit request =>
-      val preparedForm = request.userAnswers.get(SpecialMentionAdditionalInfoPage) match {
+      val preparedForm = request.userAnswers.get(AddSpecialMentionPage) match {
         case None        => form
         case Some(value) => form.fill(value)
       }
 
       val json = Json.obj(
-        "form" -> preparedForm,
-        "lrn"  -> lrn,
-        "mode" -> mode
+        "form"   -> preparedForm,
+        "mode"   -> mode,
+        "lrn"    -> lrn,
+        "radios" -> Radios.yesNo(preparedForm("value"))
       )
 
       renderer.render(template, json).map(Ok(_))
@@ -75,18 +76,19 @@ class SpecialMentionAdditionalInfoController @Inject()(
           formWithErrors => {
 
             val json = Json.obj(
-              "form" -> formWithErrors,
-              "lrn"  -> lrn,
-              "mode" -> mode
+              "form"   -> formWithErrors,
+              "mode"   -> mode,
+              "lrn"    -> lrn,
+              "radios" -> Radios.yesNo(formWithErrors("value"))
             )
 
             renderer.render(template, json).map(BadRequest(_))
           },
           value =>
             for {
-              updatedAnswers <- Future.fromTry(request.userAnswers.set(SpecialMentionAdditionalInfoPage, value))
+              updatedAnswers <- Future.fromTry(request.userAnswers.set(AddSpecialMentionPage, value))
               _              <- sessionRepository.set(updatedAnswers)
-            } yield Redirect(navigator.nextPage(SpecialMentionAdditionalInfoPage, mode, updatedAnswers))
+            } yield Redirect(navigator.nextPage(AddSpecialMentionPage, mode, updatedAnswers))
         )
   }
 }
