+++ conf/app.routes
@@ -504,22 +504,22 @@ POST       /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/extra-
 GET        /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-extra-information                 controllers.addItems.previousReferences.ExtraInformationController.onPageLoad(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 POST       /:lrn/add-items/:itemIndex/previous-references/:referenceIndex/change-extra-information                 controllers.addItems.previousReferences.ExtraInformationController.onSubmit(lrn: LocalReferenceNumber, itemIndex: Index, referenceIndex: Index, mode: Mode = CheckMode)
 
-GET        /:lrn/addSpecialMention                        controllers.AddSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/addSpecialMention                        controllers.AddSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/changeAddSpecialMention                  controllers.AddSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/changeAddSpecialMention                  controllers.AddSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/add-special-mentions                        controllers.specialMentions.AddSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+POST       /:lrn/add-special-mentions                        controllers.specialMentions.AddSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+GET        /:lrn/change-add-special-mentions                  controllers.specialMentions.AddSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+POST       /:lrn/change-add-special-mentions                  controllers.specialMentions.AddSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 
-GET        /:lrn/specialMentionType                        controllers.SpecialMentionTypeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/specialMentionType                        controllers.SpecialMentionTypeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/changeSpecialMentionType                  controllers.SpecialMentionTypeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/changeSpecialMentionType                  controllers.SpecialMentionTypeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/special-mention-type                        controllers.specialMentions.SpecialMentionTypeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+POST       /:lrn/special-mention-type                        controllers.specialMentions.SpecialMentionTypeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+GET        /:lrn/change-special-mention-type               controllers.specialMentions.SpecialMentionTypeController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+POST       /:lrn/change-special-mention-type                  controllers.specialMentions.SpecialMentionTypeController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 
-GET        /:lrn/specialMentionAdditionalInfo                        controllers.SpecialMentionAdditionalInfoController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/specialMentionAdditionalInfo                        controllers.SpecialMentionAdditionalInfoController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/changeSpecialMentionAdditionalInfo                  controllers.SpecialMentionAdditionalInfoController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/changeSpecialMentionAdditionalInfo                  controllers.SpecialMentionAdditionalInfoController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/special-mentions-additional-info                        controllers.specialMentions.SpecialMentionAdditionalInfoController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+POST       /:lrn/special-mentions-additional-info                        controllers.specialMentions.SpecialMentionAdditionalInfoController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+GET        /:lrn/change-special-mentions-additional-info                  controllers.specialMentions.SpecialMentionAdditionalInfoController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+POST       /:lrn/change-special-mentions-additional-info                  controllers.specialMentions.SpecialMentionAdditionalInfoController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
 
-GET        /:lrn/addAnotherSpecialMention                        controllers.AddAnotherSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-POST       /:lrn/addAnotherSpecialMention                        controllers.AddAnotherSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
-GET        /:lrn/changeAddAnotherSpecialMention                  controllers.AddAnotherSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
-POST       /:lrn/changeAddAnotherSpecialMention                  controllers.AddAnotherSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+GET        /:lrn/add-another-special-mention                        controllers.specialMentions.AddAnotherSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+POST       /:lrn/add-another-special-mention                        controllers.specialMentions.AddAnotherSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = NormalMode)
+GET        /:lrn/change-add-another-special-mention                  controllers.specialMentions.AddAnotherSpecialMentionController.onPageLoad(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
+POST       /:lrn/change-add-another-special-mention                  controllers.specialMentions.AddAnotherSpecialMentionController.onSubmit(lrn: LocalReferenceNumber, mode: Mode = CheckMode)
