+++ conf/views/addAnotherSpecialMention.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.AddAnotherSpecialMentionController.onSubmit(lrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.specialMentions.AddAnotherSpecialMentionController.onSubmit(lrn, mode).url }}" method="post">
 
           {{ csrf() | safe }}
