+++ conf/views/addSpecialMention.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.AddSpecialMentionController.onSubmit(lrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.specialMentions.AddSpecialMentionController.onSubmit(lrn, mode).url }}" method="post">
 
           {{ csrf() | safe }}
