+++ conf/views/specialMentionAdditionalInfo.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.SpecialMentionAdditionalInfoController.onSubmit(lrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.specialMentions.SpecialMentionAdditionalInfoController.onSubmit(lrn, mode).url }}" method="post">
 
           {{ csrf() | safe }}
