+++ conf/views/specialMentionType.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.SpecialMentionTypeController.onSubmit(lrn, mode).url }}" method="post">
+        <form action="{{ routes.controllers.specialMentions.SpecialMentionTypeController.onSubmit(lrn, mode).url }}" method="post">
 
           {{ csrf() | safe }}
