+++ test/controllers/specialMentions/AddSpecialMentionControllerSpec.scala
@@ -14,48 +14,46 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.specialMentions
 
-import base.SpecBase
-import base.MockNunjucksRendererApp
+import base.{MockNunjucksRendererApp, SpecBase}
+import forms.AddSpecialMentionFormProvider
 import matchers.JsonMatchers
-import forms.SpecialMentionAdditionalInfoFormProvider
 import models.{NormalMode, UserAnswers}
-import navigation.{FakeNavigator, Navigator}
 import navigation.annotations.PreTaskListDetails
+import navigation.{FakeNavigator, Navigator}
 import org.mockito.ArgumentCaptor
 import org.mockito.ArgumentMatchers.any
 import org.mockito.Mockito.{times, verify, when}
 import org.scalatestplus.mockito.MockitoSugar
-import pages.SpecialMentionAdditionalInfoPage
+import pages.AddSpecialMentionPage
 import play.api.inject.bind
 import play.api.inject.guice.GuiceApplicationBuilder
-import play.api.libs.json.{JsObject, JsString, Json}
+import play.api.libs.json.{JsObject, Json}
 import play.api.mvc.Call
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import play.twirl.api.Html
-import repositories.SessionRepository
-import uk.gov.hmrc.viewmodels.NunjucksSupport
+import uk.gov.hmrc.viewmodels.{NunjucksSupport, Radios}
 
 import scala.concurrent.Future
 
-class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunjucksRendererApp with MockitoSugar with NunjucksSupport with JsonMatchers {
+class AddSpecialMentionControllerSpec extends SpecBase with MockNunjucksRendererApp with MockitoSugar with NunjucksSupport with JsonMatchers {
 
   def onwardRoute = Call("GET", "/foo")
 
-  private val formProvider = new SpecialMentionAdditionalInfoFormProvider()
+  private val formProvider = new AddSpecialMentionFormProvider()
   private val form         = formProvider()
-  private val template     = "specialMentionAdditionalInfo.njk"
+  private val template     = "addSpecialMention.njk"
 
-  lazy val specialMentionAdditionalInfoRoute = routes.SpecialMentionAdditionalInfoController.onPageLoad(lrn, NormalMode).url
+  lazy val addSpecialMentionRoute = routes.AddSpecialMentionController.onPageLoad(lrn, NormalMode).url
 
   override def guiceApplicationBuilder(): GuiceApplicationBuilder =
     super
       .guiceApplicationBuilder()
       .overrides(bind(classOf[Navigator]).qualifiedWith(classOf[PreTaskListDetails]).toInstance(new FakeNavigator(onwardRoute)))
 
-  "SpecialMentionAdditionalInfo Controller" - {
+  "AddSpecialMention Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
@@ -64,7 +62,7 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
 
       dataRetrievalWithData(emptyUserAnswers)
 
-      val request        = FakeRequest(GET, specialMentionAdditionalInfoRoute)
+      val request        = FakeRequest(GET, addSpecialMentionRoute)
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
 
@@ -75,9 +73,10 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
       val expectedJson = Json.obj(
-        "form" -> form,
-        "mode" -> NormalMode,
-        "lrn"  -> lrn
+        "form"   -> form,
+        "mode"   -> NormalMode,
+        "lrn"    -> lrn,
+        "radios" -> Radios.yesNo(form("value"))
       )
 
       val jsonWithoutConfig = jsonCaptor.getValue - configKey
@@ -92,10 +91,10 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
       when(mockRenderer.render(any(), any())(any()))
         .thenReturn(Future.successful(Html("")))
 
-      val userAnswers = emptyUserAnswers.set(SpecialMentionAdditionalInfoPage, "answer").success.value
+      val userAnswers = UserAnswers(lrn, eoriNumber).set(AddSpecialMentionPage, true).success.value
       dataRetrievalWithData(userAnswers)
 
-      val request        = FakeRequest(GET, specialMentionAdditionalInfoRoute)
+      val request        = FakeRequest(GET, addSpecialMentionRoute)
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
 
@@ -105,12 +104,13 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
 
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
-      val filledForm = form.bind(Map("value" -> "answer"))
+      val filledForm = form.bind(Map("value" -> "true"))
 
       val expectedJson = Json.obj(
-        "form" -> filledForm,
-        "lrn"  -> lrn,
-        "mode" -> NormalMode
+        "form"   -> filledForm,
+        "mode"   -> NormalMode,
+        "lrn"    -> lrn,
+        "radios" -> Radios.yesNo(filledForm("value"))
       )
 
       val jsonWithoutConfig = jsonCaptor.getValue - configKey
@@ -127,12 +127,13 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
       dataRetrievalWithData(emptyUserAnswers)
 
       val request =
-        FakeRequest(POST, specialMentionAdditionalInfoRoute)
-          .withFormUrlEncodedBody(("value", "answer"))
+        FakeRequest(POST, addSpecialMentionRoute)
+          .withFormUrlEncodedBody(("value", "true"))
 
       val result = route(app, request).value
 
       status(result) mustEqual SEE_OTHER
+
       redirectLocation(result).value mustEqual onwardRoute.url
 
     }
@@ -144,7 +145,7 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
 
       dataRetrievalWithData(emptyUserAnswers)
 
-      val request        = FakeRequest(POST, specialMentionAdditionalInfoRoute).withFormUrlEncodedBody(("value", ""))
+      val request        = FakeRequest(POST, addSpecialMentionRoute).withFormUrlEncodedBody(("value", ""))
       val boundForm      = form.bind(Map("value" -> ""))
       val templateCaptor = ArgumentCaptor.forClass(classOf[String])
       val jsonCaptor     = ArgumentCaptor.forClass(classOf[JsObject])
@@ -156,13 +157,16 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
       verify(mockRenderer, times(1)).render(templateCaptor.capture(), jsonCaptor.capture())(any())
 
       val expectedJson = Json.obj(
-        "form" -> boundForm,
-        "lrn"  -> lrn,
-        "mode" -> NormalMode
+        "form"   -> boundForm,
+        "mode"   -> NormalMode,
+        "lrn"    -> lrn,
+        "radios" -> Radios.yesNo(boundForm("value"))
       )
 
+      val jsonWithoutConfig = jsonCaptor.getValue - configKey
+
       templateCaptor.getValue mustEqual template
-      jsonCaptor.getValue must containJson(expectedJson)
+      jsonWithoutConfig mustBe expectedJson
 
     }
 
@@ -170,13 +174,13 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
 
       dataRetrievalNoData()
 
-      val request = FakeRequest(GET, specialMentionAdditionalInfoRoute)
+      val request = FakeRequest(GET, addSpecialMentionRoute)
 
       val result = route(app, request).value
 
       status(result) mustEqual SEE_OTHER
 
-      redirectLocation(result).value mustEqual routes.SessionExpiredController.onPageLoad().url
+      redirectLocation(result).value mustEqual controllers.routes.SessionExpiredController.onPageLoad().url
 
     }
 
@@ -185,14 +189,14 @@ class SpecialMentionAdditionalInfoControllerSpec extends SpecBase with MockNunju
       dataRetrievalNoData()
 
       val request =
-        FakeRequest(POST, specialMentionAdditionalInfoRoute)
-          .withFormUrlEncodedBody(("value", "answer"))
+        FakeRequest(POST, addSpecialMentionRoute)
+          .withFormUrlEncodedBody(("value", "true"))
 
       val result = route(app, request).value
 
       status(result) mustEqual SEE_OTHER
 
-      redirectLocation(result).value mustEqual routes.SessionExpiredController.onPageLoad().url
+      redirectLocation(result).value mustEqual controllers.routes.SessionExpiredController.onPageLoad().url
 
     }
   }
